requirements.txt
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
idna==2.8
Pillow==7.0.0
pony==0.7.13
psycopg2==2.8.4
requests==2.22.0
six==1.13.0
urllib3==1.25.7
vk-api==11.6.1

